Add in an IDBUpgrader interface to help create tables.
I'm not sure that the interface name is right for the long term, but for now at least it's functional.  Changed the tests to use the IDBUpgrader object that's now available off of IDatabase, and table creation is now a lot faster, presumably because it no longer has to trawl around on disk re-discovering the .ddl file.