Clean up the persistence handling around columns.
Added in an IDBColumnTypePersistenceHandler interface, which can be
implemented in various ways.  Ultimately that may turn into an internal-only API that's flattened out on IDBColumnType, however.  Add in a wrapParameterValue method in IDBColumn that chains through to use the persistence handler on the column's type, and change all existing methods that use database.wrapParameter() to instead use the method off of the IDBColumn.