Introduce select() vs. selectAsStruct() distinction:
The select() method of a '*' query now returns an iterable of the root table type even in the presence of joins.

I chose this so that someone who wished to add a join to a query that did not have one previously would not change
the return type of the query.

I added a new method, selectAsStruct(), in cases where there are joins, that give you a strongly typed struct of the
fields of all tables involved (although in conflicting columns, the results are not specified, it will be driver
dependent.)