Support an OPTIONAL() boolean expression
This is an experimental, small deviation from the SQL99 grammar.  I hope the expressivity increase turns out to be worth the cost.

This expression form allows users to create queries that change form based on what values are passed in.  For example,
consider this query:

  SELECT *
  FROM FOO
  WHERE
    OPTIONAL( COL1 > :var1 ) AND
    OPTIONAL( COL2 < :var2 ) AND
    OPTIONAL( COL3 = :var3 )

The OPTIONAL clauses will only be included in the query if all the variables mentioned within the OPTIONAL have non-null values.

This supports a common pattern where search-like functionality can include multiple conditions based on the values that a user passes in.

Interesting note on OPTION and OR: I've tried to keep the semantics of OR as intuitive as possible.  If an OR has OPTION expression on either side, and if neither of those OPTION expressions are going to apply, we erase the OR expression from the entire boolean clause.

So, given this query:

  SELECT *
  FROM FOO
  WHERE
    OPTIONAL( COL1 > :var1 ) OR
    OPTIONAL( COL2 < :var2 ) OR
    OPTIONAL( COL3 = :var3 )

If :var1 and :var2 are null, the resulting query will be:

  SELECT *
  FROM FOO
  WHERE
    COL3 = :var3

which I think is the intuitive behavior.