Support comments in tokenizer and in .sql files
I only support line comments of the form '--' right now, I'll introduce
C-style comments later.  Currently comments are emitted as a token with
a new TokenType: COMMENT.  I also introduced a way to remove tokens of
a certain kind, which I use to remove comments before I parse the SQL.

The DDL parser may wish to remove them or to keep them as part of the grammar.