Add in a select(String, Map) method to all db types.
The select method takes a String that must start with SELECT * FROM
and returns a QueryResult of objects.  The second parameter is optional,
and if present will be used to do argument substitution on the sql
string passed in.  Tokens of the form :arg will be substituted with
the value in the map for "arg" and will appear in the generated sql
as prepared statement parameters.