Move most CachedDBObject implementation into Gosu.
As I found out the hard way, the entire implementation can't be in Gosu, since
Gosu automatically implements getIntrinsicType() on Gosu objects, and we need
to override that to return the appropriate Tosa type.  So instead, the Java
code now delegates out to a Gosu implementation, which contains all the
actual logic.